requirements.txt.txt
# Core dependencies
Python >=3.10
numpy
torch
pyyaml
opencv-python
tqdm
openai-whisper
googletrans
openvino-dev
intel-extension-for-pytorch
ultralytics  
PySide6      
ffmpeg





